+++ django-dynamic-map-borinud/dynamic/templates/base_map.html
@@ -81,6 +81,8 @@
         crossorigin="anonymous"></script>
 <script src="{% static 'dynamic/js/html2canvas.min.js' %}"></script>
 <script src="{% static 'dynamic/js/print.min.js' %}"></script>
+<script src="{% static 'dynamic/js/wms-capabilities.min.js' %}"></script>
+<script src="{% static 'dynamic/js/iso8601.min.js' %}"></script>
 {% block javascript %}
 {% endblock %}
 <script type="text/javascript" src="{% static 'dynamic/js/borinud.trange.js' %}"></script>
